pyproject.toml
[project]
name = "advanced-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "firecrawl-py>=2.16.5",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "reflex>=0.8.6",
]
